🐳 😃 Reorganize metrics and expectation functions (#763)
* Cleanup expectation functions

Trigger CI

* Add helper function

Trigger CI
